requirements.txt
# Automation packages
behave
selenium
webdriver-manager
allure-behave
undetected-chromedriver

# API packages
requests
locust

# Helpers
python-dotenv

# Initial Python and pip packages
parse
parse-type
six
urllib3